feat(backgroundImage): background image component creation
