+++ bluebase/common/bluebase.ts
@@ -1,8 +1,4 @@
 // This file contain all the apps, plugins and configuration which are required
 // for booting bluebase. see https://blueeast.gitbooks.io/bluebase/
 export default {
-
-    assets: {
-        Wallpaper: require("../../assets/common/logo.png")
-    }
 };
